pyproject.toml
[tool.poetry]
name = "langchain-hello-world"
version = "0.1.0"
description = ""
authors = ["Naiane Rosa <naiane.rosa@valory.xyz>"]

[tool.poetry.dependencies]
python = ">=3.9,<4.0"
langchain-community = "^0.3.16"
langgraph = ">=0.2.66,<0.3.0"
langchain-anthropic = ">=0.3.3,<0.4.0"
tavily-python = ">=0.5.0,<0.6.0"
langgraph-checkpoint-sqlite = ">=2.0.3,<3.0.0"
langchain-openai = ">=0.3.1,<0.4.0"
python-dotenv = ">=1.0.1,<2.0.0"

[project]
name = "langchain-hello-world"
version = "0.1.0"
description = ""
authors = [
    {name = "Naiane Rosa",email = "naiane.rosa@valory.xyz"}
]
readme = "README.md"
requires-python = ">=3.9,<4.0"
dependencies = [
    "langchain-community (>=0.3.15,<0.4.0)",
    "langgraph (>=0.2.66,<0.3.0)",
    "langchain-anthropic (>=0.3.3,<0.4.0)",
    "tavily-python (>=0.5.0,<0.6.0)",
    "langgraph-checkpoint-sqlite (>=2.0.3,<3.0.0)",
    "langchain-openai (>=0.3.1,<0.4.0)",
    "python-dotenv (>=1.0.1,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
